requirements.txt
alembic==1.7.4
apifairy==0.8.2
apispec==5.1.1
arabic-reshaper==2.1.3
bleach==4.1.0
Brotli==1.0.9
cffi==1.15.0
click==8.0.3
cssselect2==0.4.1
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fonttools==4.27.1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
importlib-metadata==4.8.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
packaging==21.2
Pillow==8.4.0
pycparser==2.20
pydyf==0.1.2
PyMySQL==1.0.2
pyparsing==2.4.7
pyphen==0.11.0
python-bidi==0.4.2
python-dotenv==0.19.1
reportlab==3.6.2
six==1.16.0
SQLAlchemy==1.4.26
tinycss2==1.1.0
weasyprint==53.3
webargs==8.0.1
webencodings==0.5.1
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.6.0
zopfli==0.1.9
